Refactoring of the schematic execution engine in the tests
