+++ schematics/src/tests/utils/schematic.schema.ts
@@ -1,0 +1,5 @@
+export interface SchematicSchema {
+    type: string;
+    options: { [key: string]: unknown };
+    children: SchematicSchema[];
+}
